pyproject.toml
[project]
name = "polytraderai"
version = "0.1.0"
description = "Polytrade AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cryptography==43.0.3",
    "fastapi>=0.115.12",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema-rs==0.29.1",
    "langchain-core==0.3.48",
    "langchain-fireworks>=0.2.8",
    "langchain-google-genai>=2.1.1",
    "langgraph==0.3.19",
    "langgraph-api==0.0.33",
    "langgraph-checkpoint==2.0.23",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langgraph-cli==0.1.79",
    "langgraph-prebuilt==0.1.4",
    "langgraph-sdk==0.1.58",
    "langsmith==0.3.18",
    "msgpack==1.1.0",
    "newsapi-python>=0.2.7",
    "open-deep-research>=0.0.10",
    "orjson==3.10.16",
    "ormsgpack==1.9.0",
    "packaging==24.2",
    "pip-review==1.3.0",
    "py-clob-client>=0.20.0",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
    "pytrends>=4.9.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sse-starlette==2.1.3",
    "starlette==0.46.1",
    "structlog==25.2.0",
    "tenacity==9.0.0",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "web3>=7.9.0",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
